+++ Data/Posicion.h
@@ -1,19 +1,19 @@
 #pragma once
 class Posicion
 {
-	int x, y;
+	float x, y;
 public:
 	Posicion();
-	Posicion(int x);
-	Posicion(int x, int y);
+	Posicion(float x);
+	Posicion(float x, float y);
 	Posicion(const Posicion& pos);
 	~Posicion();
 
-	void Set(int x, int y);
+	void Set(float x, float y);
 	void Set(const Posicion& pos);
 	Posicion& Get();
-	int GetX() const;
-	int GetY() const;
+	float GetX() const;
+	float GetY() const;
 
 	Posicion EnTiles();
 };
